#440: Refactor unit tests and split according to cases.
