+++ src/test/java/com/commercetools/sync/products/helpers/variantreferenceresolver/WithProductReferencesTest.java
@@ -9,11 +9,10 @@ import com.commercetools.sync.services.CustomerGroupService;
 import com.commercetools.sync.services.ProductService;
 import com.commercetools.sync.services.TypeService;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import io.sphere.sdk.categories.Category;
 import io.sphere.sdk.client.SphereClient;
+import io.sphere.sdk.products.Product;
 import io.sphere.sdk.products.ProductVariantDraft;
 import io.sphere.sdk.products.ProductVariantDraftBuilder;
 import io.sphere.sdk.products.attributes.AttributeDraft;
@@ -28,8 +27,9 @@ import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 
-import static com.commercetools.sync.products.ProductSyncMockUtils.createReferenceObject;
-import static com.commercetools.sync.products.ProductSyncMockUtils.getMockCategoryService;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getMockProductService;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getProductReferenceWithId;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getProductReferenceWithRandomId;
 import static com.commercetools.sync.products.ProductSyncMockUtils.getReferenceSetAttributeDraft;
 import static com.commercetools.sync.products.helpers.VariantReferenceResolver.REFERENCE_ID_FIELD;
 import static com.commercetools.sync.products.helpers.VariantReferenceResolver.REFERENCE_TYPE_ID_FIELD;
@@ -38,10 +38,9 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-class VariantReferenceResolverWithCategoryReferencesTest {
-    private CategoryService categoryService;
-    private static final String CATEGORY1_ID = UUID.randomUUID().toString();
-    private static final String CATEGORY2_ID = UUID.randomUUID().toString();
+class WithProductReferencesTest {
+    private ProductService productService;
+    private static final String PRODUCT_ID = UUID.randomUUID().toString();
     private VariantReferenceResolver referenceResolver;
 
     /**
@@ -49,23 +48,22 @@ class VariantReferenceResolverWithCategoryReferencesTest {
      */
     @BeforeEach
     void setup() {
-        categoryService = getMockCategoryService(CATEGORY1_ID, CATEGORY2_ID);
+        productService = getMockProductService(PRODUCT_ID);
         final ProductSyncOptions syncOptions = ProductSyncOptionsBuilder.of(mock(SphereClient.class)).build();
-        referenceResolver = new VariantReferenceResolver(syncOptions,
-            mock(TypeService.class),
+        referenceResolver = new VariantReferenceResolver(syncOptions, mock(TypeService.class),
             mock(ChannelService.class),
             mock(CustomerGroupService.class),
-            mock(ProductService.class),
-            categoryService);
+            productService,
+            mock(CategoryService.class));
     }
 
     @Test
-    void resolveReferences_WithNonExistingCategoryReferenceAttribute_ShouldReturnEqualDraft() {
+    void resolveReferences_WithNonExistingProductReferenceAttribute_ShouldReturnEqualDraft() {
         // preparation
-        when(categoryService.fetchCachedCategoryId(anyString()))
+        when(productService.getIdFromCacheOrFetch(anyString()))
             .thenReturn(CompletableFuture.completedFuture(Optional.empty()));
 
-        final ObjectNode attributeValue = createReferenceObject("nonExistingCatKey", Category.referenceTypeId());
+        final ObjectNode attributeValue = getProductReferenceWithRandomId();
         final AttributeDraft attributeDraft = AttributeDraft.of("attributeName", attributeValue);
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
@@ -82,14 +80,14 @@ class VariantReferenceResolverWithCategoryReferencesTest {
     }
 
     @Test
-    void resolveReferences_WithNullIdFieldInCategoryReferenceAttribute_ShouldReturnEqualDraft() {
+    void resolveReferences_WithNullIdFieldInProductReferenceAttribute_ShouldReturnEqualDraft() {
         // preparation
         final ObjectNode attributeValue = JsonNodeFactory.instance.objectNode();
-        attributeValue.put(REFERENCE_TYPE_ID_FIELD, Category.referenceTypeId());
-        final AttributeDraft categoryReferenceAttribute = AttributeDraft.of("attributeName", attributeValue);
+        attributeValue.put(REFERENCE_TYPE_ID_FIELD, Product.referenceTypeId());
+        final AttributeDraft productReferenceAttribute = AttributeDraft.of("attributeName", attributeValue);
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceAttribute)
             .build();
 
         // test
@@ -102,15 +100,15 @@ class VariantReferenceResolverWithCategoryReferencesTest {
     }
 
     @Test
-    void resolveReferences_WithNullNodeIdFieldInCategoryReferenceAttribute_ShouldReturnEqualDraft() {
+    void resolveReferences_WithNullNodeIdFieldInProductReferenceAttribute_ShouldReturnEqualDraft() {
         // preparation
         final ObjectNode attributeValue = JsonNodeFactory.instance.objectNode();
-        attributeValue.put(REFERENCE_TYPE_ID_FIELD, Category.referenceTypeId());
-        final AttributeDraft categoryReferenceAttribute =
+        attributeValue.put(REFERENCE_TYPE_ID_FIELD, Product.referenceTypeId());
+        final AttributeDraft productReferenceAttribute =
             AttributeDraft.of("attributeName", JsonNodeFactory.instance.nullNode());
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceAttribute)
             .build();
 
         // test
@@ -123,13 +121,13 @@ class VariantReferenceResolverWithCategoryReferencesTest {
     }
 
     @Test
-    void resolveReferences_WithExistingCategoryReferenceAttribute_ShouldReturnResolvedDraft() {
+    void resolveReferences_WithExistingProductReferenceAttribute_ShouldReturnResolvedDraft() {
         // preparation
-        final ObjectNode attributeValue = createReferenceObject("foo", Category.referenceTypeId());
-        final AttributeDraft categoryReferenceAttribute = AttributeDraft.of("attributeName", attributeValue);
+        final ObjectNode attributeValue = getProductReferenceWithRandomId();
+        final AttributeDraft productReferenceAttribute = AttributeDraft.of("attributeName", attributeValue);
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceAttribute)
             .build();
 
         // test
@@ -141,22 +139,20 @@ class VariantReferenceResolverWithCategoryReferencesTest {
         assertThat(resolvedAttributeDraft.getAttributes()).isNotNull();
         final AttributeDraft resolvedAttribute = resolvedAttributeDraft.getAttributes().get(0);
         assertThat(resolvedAttribute).isNotNull();
-        assertThat(resolvedAttribute.getValue().get(REFERENCE_ID_FIELD).asText()).isEqualTo(CATEGORY1_ID);
+        assertThat(resolvedAttribute.getValue().get(REFERENCE_ID_FIELD).asText()).isEqualTo(PRODUCT_ID);
         assertThat(resolvedAttribute.getValue().get(REFERENCE_TYPE_ID_FIELD).asText())
-            .isEqualTo(Category.referenceTypeId());
+            .isEqualTo(Product.referenceTypeId());
     }
 
     @Test
-    void resolveReferences_WithCategoryReferenceSetAttribute_ShouldResolveReferences() {
-        // preparation
-        final ArrayNode referenceSet = JsonNodeFactory.instance.arrayNode();
-        referenceSet.add(createReferenceObject("foo", Category.referenceTypeId()));
-        referenceSet.add(createReferenceObject("bar", Category.referenceTypeId()));
-        final AttributeDraft categoryReferenceAttribute = AttributeDraft.of("attributeName", referenceSet);
+    void resolveReferences_WithProductReferenceSetAttribute_ShouldResolveReferences() {
+        final ObjectNode productReferenceWithRandomId = getProductReferenceWithRandomId();
+        final AttributeDraft productReferenceSetAttributeDraft =
+            getReferenceSetAttributeDraft("foo", productReferenceWithRandomId);
 
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceSetAttributeDraft)
             .build();
 
         // test
@@ -168,9 +164,7 @@ class VariantReferenceResolverWithCategoryReferencesTest {
         assertThat(resolvedProductVariantDraft).isNotNull();
         assertThat(resolvedProductVariantDraft.getAttributes()).isNotNull();
 
-
         final AttributeDraft resolvedAttributeDraft = resolvedProductVariantDraft.getAttributes().get(0);
-
         assertThat(resolvedAttributeDraft).isNotNull();
         assertThat(resolvedAttributeDraft.getValue()).isNotNull();
 
@@ -179,25 +173,25 @@ class VariantReferenceResolverWithCategoryReferencesTest {
         final Set<JsonNode> resolvedSet = StreamSupport.stream(attributeReferencesIterator, false)
                                                        .collect(Collectors.toSet());
         assertThat(resolvedSet).isNotEmpty();
-        assertThat(resolvedSet).containsExactlyInAnyOrder(
-            createReferenceObject(CATEGORY1_ID, Category.referenceTypeId()),
-            createReferenceObject(CATEGORY2_ID, Category.referenceTypeId()));
+        final ObjectNode resolvedReference = JsonNodeFactory.instance.objectNode();
+        resolvedReference.put(REFERENCE_TYPE_ID_FIELD, "product");
+        resolvedReference.put(REFERENCE_ID_FIELD, PRODUCT_ID);
+        assertThat(resolvedSet).containsExactly(resolvedReference);
     }
 
     @Test
-    void resolveReferences_WithNonExistingCategoryReferenceSetAttribute_ShouldNotResolveReferences() {
+    void resolveReferences_WithNonExistingProductReferenceSetAttribute_ShouldNotResolveReferences() {
         // preparation
-        when(categoryService.fetchCachedCategoryId(anyString()))
+        when(productService.getIdFromCacheOrFetch(anyString()))
             .thenReturn(CompletableFuture.completedFuture(Optional.empty()));
 
-        final ObjectNode categoryReference =
-            createReferenceObject(UUID.randomUUID().toString(), Category.referenceTypeId());
-        final AttributeDraft categoryReferenceAttribute =
-            getReferenceSetAttributeDraft("foo", categoryReference);
+        final ObjectNode productReference = getProductReferenceWithRandomId();
+        final AttributeDraft productReferenceAttribute =
+            getReferenceSetAttributeDraft("foo", productReference);
 
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceAttribute)
             .build();
 
         // test
@@ -218,26 +212,26 @@ class VariantReferenceResolverWithCategoryReferencesTest {
         assertThat(attributeReferencesIterator).isNotNull();
         final Set<JsonNode> resolvedSet = StreamSupport.stream(attributeReferencesIterator, false)
                                                        .collect(Collectors.toSet());
-        assertThat(resolvedSet).containsExactly(categoryReference);
+        assertThat(resolvedSet).containsExactly(productReference);
     }
 
     @Test
-    void resolveReferences_WithSomeExistingCategoryReferenceSetAttribute_ShouldResolveExistingReferences() {
+    void resolveReferences_WithSomeExistingProductReferenceSetAttribute_ShouldResolveExistingReferences() {
         // preparation
-        when(categoryService.fetchCachedCategoryId("existingKey"))
+        when(productService.getIdFromCacheOrFetch("existingKey"))
             .thenReturn(CompletableFuture.completedFuture(Optional.of("existingId")));
-        when(categoryService.fetchCachedCategoryId("randomKey"))
+        when(productService.getIdFromCacheOrFetch("randomKey"))
             .thenReturn(CompletableFuture.completedFuture(Optional.empty()));
 
-        final ObjectNode categoryReference1 = createReferenceObject("existingKey", Category.referenceTypeId());
-        final ObjectNode categoryReference2 = createReferenceObject("randomKey", Category.referenceTypeId());
+        final ObjectNode productReference1 = getProductReferenceWithId("existingKey");
+        final ObjectNode productReference2 = getProductReferenceWithId("randomKey");
 
-        final AttributeDraft categoryReferenceAttribute =
-            getReferenceSetAttributeDraft("foo", categoryReference1, categoryReference2);
+        final AttributeDraft productReferenceAttribute =
+            getReferenceSetAttributeDraft("foo", productReference1, productReference2);
 
         final ProductVariantDraft productVariantDraft = ProductVariantDraftBuilder
             .of()
-            .attributes(categoryReferenceAttribute)
+            .attributes(productReferenceAttribute)
             .build();
 
         // test
@@ -259,8 +253,8 @@ class VariantReferenceResolverWithCategoryReferencesTest {
         final Set<JsonNode> resolvedSet = StreamSupport.stream(attributeReferencesIterator, false)
                                                        .collect(Collectors.toSet());
 
-        final ObjectNode resolvedReference1 = createReferenceObject("existingId", Category.referenceTypeId());
-        final ObjectNode resolvedReference2 = createReferenceObject("randomKey", Category.referenceTypeId());
+        final ObjectNode resolvedReference1 = getProductReferenceWithId("existingId");
+        final ObjectNode resolvedReference2 = getProductReferenceWithId("randomKey");
         assertThat(resolvedSet).containsExactlyInAnyOrder(resolvedReference1, resolvedReference2);
     }
 }
