+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withnestedattributes/withcategoryreferences/with-non-existing-references.json
@@ -1,0 +1,33 @@
+{
+  "key": "key",
+  "sku": "sku",
+  "attributes": [
+    {
+      "name": "attribute-name",
+      "value": [
+        {
+          "value": {
+            "id": "nonExistingCategoryKey1",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-1-name"
+        },
+        {
+          "value": {
+            "id": "key1",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-2-name"
+        },
+        {
+          "value": {
+            "id": "nonExistingCategoryKey3",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-3-name"
+        }
+      ]
+    }
+  ]
+}
+
