+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withnestedattributes/withcategoryreferences/with-reference.json
@@ -1,0 +1,40 @@
+{
+  "key": "key",
+  "sku": "sku",
+  "attributes": [
+    {
+      "name": "attribute-name",
+      "value": [
+        {
+          "value": {
+            "id": "key1",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-1-name"
+        },
+        {
+          "value": {
+            "id": "key2",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-2-name"
+        },
+        {
+          "value": {
+            "id": "key3",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-3-name"
+        },
+        {
+          "value": {
+            "id": "key4",
+            "typeId": "category"
+          },
+          "name": "nested-attribute-4-name"
+        }
+      ]
+    }
+  ]
+}
+
