+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withnestedattributes/withnoreferences/with-set-of-text-attributes.json
