+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withnestedattributes/withnoreferences/with-text-attributes.json
