+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withnestedattributes/withproductreferences/with-reference.json
