+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withsetofnestedattributes/withnoreferences/with-set-of-text-attributes.json
@@ -1,0 +1,40 @@
+{
+  "key": "key",
+  "sku": "sku",
+  "attributes": [
+    {
+      "name": "attribute-name",
+      "value": [
+        [
+          {
+            "value": [
+              "nested-attribute-1-value-a",
+              "nested-attribute-1-value-b"
+            ],
+            "name": "nested-attribute-1-name"
+          },
+          {
+            "value": [
+              "nested-attribute-2-value-a",
+              "nested-attribute-2-value-b"
+            ],
+            "name": "nested-attribute-2-name"
+          },
+          {
+            "value": [
+              "nested-attribute-3-value"
+            ],
+            "name": "nested-attribute-3-name"
+          },
+          {
+            "value": [
+              "nested-attribute-4-value-a",
+              "nested-attribute-4-value-b"
+            ],
+            "name": "nested-attribute-4-name"
+          }
+        ]
+      ]
+    }
+  ]
+}
