+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withsetofnestedattributes/withnoreferences/with-text-attributes.json
@@ -1,0 +1,29 @@
+{
+  "key": "key",
+  "sku": "sku",
+  "attributes": [
+    {
+      "name": "attribute-name",
+      "value": [
+        [
+          {
+            "value": "nested-attribute-1-value",
+            "name": "nested-attribute-1-name"
+          },
+          {
+            "value": "nested-attribute-2-value",
+            "name": "nested-attribute-2-name"
+          },
+          {
+            "value": "nested-attribute-3-value",
+            "name": "nested-attribute-3-name"
+          },
+          {
+            "value": "nested-attribute-4-value",
+            "name": "nested-attribute-4-name"
+          }
+        ]
+      ]
+    }
+  ]
+}
