+++ src/test/resources/com/commercetools/sync/products/helpers/variantReferenceResolver/withsetofnestedattributes/withproductreferences/with-set-of-references.json
@@ -1,0 +1,48 @@
+{
+  "key": "key",
+  "sku": "sku",
+  "attributes": [
+    {
+      "name": "attribute-name",
+      "value": [
+        [
+          {
+            "value": [
+              {
+                "id": "key1",
+                "typeId": "product"
+              },
+              {
+                "id": "key2",
+                "typeId": "product"
+              }
+            ],
+            "name": "nested-attribute-1-name"
+          },
+          {
+            "value": {
+              "id": "key2",
+              "typeId": "product"
+            },
+            "name": "nested-attribute-2-name"
+          },
+          {
+            "value": {
+              "id": "key3",
+              "typeId": "product"
+            },
+            "name": "nested-attribute-3-name"
+          },
+          {
+            "value": {
+              "id": "key4",
+              "typeId": "product"
+            },
+            "name": "nested-attribute-4-name"
+          }
+        ]
+      ]
+    }
+  ]
+}
+
